added relative memory instructions, branch + link

+++ memory.c
@@ -10,7 +10,7 @@ struct vector *vectorTbl = (struct vector *)&(memory.word_mem[VECTORBASE>>1]);
 void bus(unsigned short MAR, unsigned short *MBR, direction dir, wb size){
   switch(dir){
   case RD:
-    *MBR = (size == W) ? memory.word_mem[MAR >> 1] : memory.byte_mem[MAR];
+    *MBR = (size == W) ? memory.word_mem[MAR >> 1] : (unsigned short)memory.byte_mem[MAR];
     break;
   case WR:
     if (size == W){
